#729: Split up components
